# src/main/resources/hmlFhirConverter-swagger-spec.yaml
swagger: '2.0'
info:
  version: "0.0.1"
  title: service-hmlFhirConverter
  description: Mid level API to handle services of web-hmlFhirAngularClient
schemes:
  - https
basePath: /v1
produces:
  - application/json
paths:
...
  /collectionMethod:
    post:
      operationId: createCollectionMethods
      description: Create testing type name entries
      parameters:
        -
          name: collection Methods
          in: body
          description: Array of Collection Method Objects
          required: true
          schema:
            title: ArrayOfCollectionMethods
            type: array
            items:
              $ref: '#/definitions/CollectionMethod'
      responses:
        200:
          description: Successful response
          schema:
            title: ArrayOfCollectionMethods
            type: array
            items:
              $ref: '#/definitions/CollectionMethod'
        500:
          description: Internal Server Error
          schema:
            title: CollectionMethodControllerException
            schema:
              $ref: '#/definitions/Error'
    put:
      operationId: updateCollectionMethod
      description: Update a single typing test name entry
      parameters:
        -
          name: collection Method
          in: body
          description: Collection Method Object
          required: true
          schema:
            $ref: '#/definitions/CollectionMethod'
      responses:
        200:
          description: Successful response
          schema:
              $ref: '#/definitions/CollectionMethod'
        500:
          description: Internal Server Error
          schema:
            title: CollectionMethodControllerException
            schema:
              $ref: '#/definitions/Error'
    delete:
      operationId: deleteCollectionMethod
      description: Delete a single typing test name entry
      parameters:
        -
          name: collection Method
          in: body
          description: Collection Method Object
          required: true
          schema:
            $ref: '#/definitions/CollectionMethod'
      responses:
        200:
          description: Successful response
          schema:
            title: BooleanOfSuccessfulOperation
            type: boolean
        500:
          description: Internal Server Error
          schema:
            title: CollectionMethodControllerException
            schema:
              $ref: '#/definitions/Error'
    get:
      operationId: getModel
      description: Get a newly instantiated model
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/CollectionMethod'
        500:
          description:
            schema:
              title: CollectionMethodControllerException
              schema:
                $ref: '#/definitions/Error'
...
  /sample:
    post:
      operationId: createSamples
      description: Create Sample entries
      parameters:
        -
          name: samples
          in: body
          description: Array of Sample Objects
          required: true
          schema:
            title: ArrayOfsamples
            type: array
            items:
              $ref: '#/definitions/Sample'
      responses:
        200:
          description: Successful response
          schema:
            title: ArrayOfsamples
            type: array
            items:
              $ref: '#/definitions/Sample'
        500:
          description: Internal Server Error
          schema:
            title: SampleControllerException
            schema:
              $ref: '#/definitions/Error'
    put:
      operationId: updateSample
      description: Update a single Sample entry
      parameters:
        -
          name: sample
          in: body
          description: Sample Object
          required: true
          schema:
            $ref: '#/definitions/Sample'
      responses:
        200:
          description: Successful response
          schema:
              $ref: '#/definitions/Sample'
        500:
          description: Internal Server Error
          schema:
            title: SampleControllerException
            schema:
              $ref: '#/definitions/Error'
    delete:
      operationId: deleteSample
      description: Delete a single Sample entry
      parameters:
        -
          name: sample
          in: body
          description: Sample Object
          required: true
          schema:
            $ref: '#/definitions/Sample'
      responses:
        200:
          description: Successful response
          schema:
            title: BooleanOfSuccessfulOperation
            type: boolean
        500:
          description: Internal Server Error
          schema:
            title: SampleControllerException
            schema:
              $ref: '#/definitions/Error'
    get:
      operationId: getModel
      description: Get a newly instantiated model
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/Sample'
        500:
          description:
            schema:
              title: SampleControllerException
              schema:
                $ref: '#/definitions/Error'
...
  /reportingCenter:
    post:
      operationId: createReportingCenters
      description: Create multiple testing type name entries
      parameters:
        -
          name: reportingCenters
          in: body
          description: Array of Reporting Center Objects
          required: true
          schema:
            title: ArrayOfreportingCenters
            type: array
            items:
              $ref: '#/definitions/ReportingCenter'
      responses:
        200:
          description: Successful response
          schema:
            title: ArrayOfreportingCenters
            type: array
            items:
              $ref: '#/definitions/ReportingCenter'
        500:
          description: Internal Server Error
          schema:
            title: ReportingCenterControllerException
            schema:
              $ref: '#/definitions/Error'
    put:
      operationId: updateReportingCenter
      description: Update a single Reporting Center entry
      parameters:
        -
          name: reportingCenter
          in: body
          description: Reporting Center Object
          required: true
          schema:
            $ref: '#/definitions/ReportingCenter'
      responses:
        200:
          description: Successful response
          schema:
              $ref: '#/definitions/ReportingCenter'
        500:
          description: Internal Server Error
          schema:
            title: ReportingCenterControllerException
            schema:
              $ref: '#/definitions/Error'
    delete:
      operationId: deleteReportingCenter
      description: Delete a single Reporting Center entry
      parameters:
        -
          name: reportingCenter
          in: body
          description: Reporting Center Object
          required: true
          schema:
            $ref: '#/definitions/ReportingCenter'
      responses:
        200:
          description: Successful response
          schema:
            title: BooleanOfSuccessfulOperation
            type: boolean
        500:
          description: Internal Server Error
          schema:
            title: ReportingCenterControllerException
            schema:
              $ref: '#/definitions/Error'
    get:
      operationId: getModel
      description: Get a newly instantiated model
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ReportingCenter'
        500:
          description:
            schema:
              title: ReportingCenterControllerException
              schema:
                $ref: '#/definitions/Error'
...
  /property:
    post:
      operationId: createProperties
      description: Create Property entries
      parameters:
        -
          name: properties
          in: body
          description: Array of Property Objects
          required: true
          schema:
            title: ArrayOfproperties
            type: array
            items:
              $ref: '#/definitions/Property'
      responses:
        200:
          description: Successful response
          schema:
            title: ArrayOfproperties
            type: array
            items:
              $ref: '#/definitions/Property'
        500:
          description: Internal Server Error
          schema:
            title: PropertyControllerException
            schema:
              $ref: '#/definitions/Error'
    put:
      operationId: updateProperty
      description: Update a single Property entry
      parameters:
        -
          name: property
          in: body
          description: Property Object
          required: true
          schema:
            $ref: '#/definitions/Property'
      responses:
        200:
          description: Successful response
          schema:
              $ref: '#/definitions/Property'
        500:
          description: Internal Server Error
          schema:
            title: PropertyControllerException
            schema:
              $ref: '#/definitions/Error'
    delete:
      operationId: deleteProperty
      description: Delete a single Property entry
      parameters:
        -
          name: property
          in: body
          description: Property Object
          required: true
          schema:
            $ref: '#/definitions/Property'
      responses:
        200:
          description: Successful response
          schema:
            title: BooleanOfSuccessfulOperation
            type: boolean
        500:
          description: Internal Server Error
          schema:
            title: PropertyControllerException
            schema:
              $ref: '#/definitions/Error'
    get:
      operationId: getModel
      description: Get a newly instantiated model
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/Property'
        500:
          description:
            schema:
              title: PropertyControllerException
              schema:
                $ref: '#/definitions/Error'
...
  /hml:
    post:
      operationId: createHmls
      description: Create Hml entries
      parameters:
        -
          name: hmls
          in: body
          description: Array of Hml Objects
          required: true
          schema:
            title: ArrayOfhmls
            type: array
            items:
              $ref: '#/definitions/Hml'
      responses:
        200:
          description: Successful response
          schema:
            title: ArrayOfhmls
            type: array
            items:
              $ref: '#/definitions/Hml'
        500:
          description: Internal Server Error
          schema:
            title: HmlControllerException
            schema:
              $ref: '#/definitions/Error'
    put:
      operationId: updateHml
      description: Update a single Hml entry
      parameters:
        -
          name: hml
          in: body
          description: Hml Object
          required: true
          schema:
            $ref: '#/definitions/Hml'
      responses:
        200:
          description: Successful response
          schema:
              $ref: '#/definitions/Hml'
        500:
          description: Internal Server Error
          schema:
            title: HmlControllerException
            schema:
              $ref: '#/definitions/Error'
    delete:
      operationId: deleteHml
      description: Delete a single Hml entry
      parameters:
        -
          name: hml
          in: body
          description: Hml Object
          required: true
          schema:
            $ref: '#/definitions/Hml'
      responses:
        200:
          description: Successful response
          schema:
            title: BooleanOfSuccessfulOperation
            type: boolean
        500:
          description: Internal Server Error
          schema:
            title: HmlControllerException
            schema:
              $ref: '#/definitions/Error'
    get:
      operationId: getModel
      description: Get a newly instantiated model
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/Hml'
        500:
          description:
            schema:
              title: HmlControllerException
              schema:
                $ref: '#/definitions/Error'
...
  /hmlId:
    post:
      operationId: createHmlIds
      description: Create HmlId entries
      parameters:
        -
          name: hmlIds
          in: body
          description: Array of HmlId Objects
          required: true
          schema:
            title: ArrayOfhmlIds
            type: array
            items:
              $ref: '#/definitions/HmlId'
      responses:
        200:
          description: Successful response
          schema:
            title: ArrayOfhmlIds
            type: array
            items:
              $ref: '#/definitions/HmlId'
        500:
          description: Internal Server Error
          schema:
            title: HmlIdControllerException
            schema:
              $ref: '#/definitions/Error'
    put:
      operationId: updateHmlId
      description: Update a single HmlId entry
      parameters:
        -
          name: hmlId
          in: body
          description: HmlId Object
          required: true
          schema:
            $ref: '#/definitions/HmlId'
      responses:
        200:
          description: Successful response
          schema:
              $ref: '#/definitions/HmlId'
        500:
          description: Internal Server Error
          schema:
            title: HmlIdControllerException
            schema:
              $ref: '#/definitions/Error'
    delete:
      operationId: deleteHmlId
      description: Delete a single HmlId entry
      parameters:
        -
          name: hmlId
          in: body
          description: HmlId Object
          required: true
          schema:
            $ref: '#/definitions/HmlId'
      responses:
        200:
          description: Successful response
          schema:
            title: BooleanOfSuccessfulOperation
            type: boolean
        500:
          description: Internal Server Error
          schema:
            title: HmlIdControllerException
            schema:
              $ref: '#/definitions/Error'
    get:
      operationId: getModel
      description: Get a newly instantiated model
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/HmlId'
        500:
          description:
            schema:
              title: HmlIdControllerException
              schema:
                $ref: '#/definitions/Error'
...
  /project:
    post:
      operationId: createProjects
      description: Create Project entries
      parameters:
        -
          name: projects
          in: body
          description: Array of Project Objects
          required: true
          schema:
            title: ArrayOfprojects
            type: array
            items:
              $ref: '#/definitions/Project'
      responses:
        200:
          description: Successful response
          schema:
            title: ArrayOfprojects
            type: array
            items:
              $ref: '#/definitions/Project'
        500:
          description: Internal Server Error
          schema:
            title: ProjectControllerException
            schema:
              $ref: '#/definitions/Error'
    put:
      operationId: updateProject
      description: Update a single Project entry
      parameters:
        -
          name: project
          in: body
          description: Project Object
          required: true
          schema:
            $ref: '#/definitions/Project'
      responses:
        200:
          description: Successful response
          schema:
              $ref: '#/definitions/Project'
        500:
          description: Internal Server Error
          schema:
            title: ProjectControllerException
            schema:
              $ref: '#/definitions/Error'
    delete:
      operationId: deleteProject
      description: Delete a single Project entry
      parameters:
        -
          name: project
          in: body
          description: Project Object
          required: true
          schema:
            $ref: '#/definitions/Project'
      responses:
        200:
          description: Successful response
          schema:
            title: BooleanOfSuccessfulOperation
            type: boolean
        500:
          description: Internal Server Error
          schema:
            title: ProjectControllerException
            schema:
              $ref: '#/definitions/Error'
    get:
      operationId: getModel
      description: Get a newly instantiated model
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/Project'
        500:
          description:
            schema:
              title: ProjectControllerException
              schema:
                $ref: '#/definitions/Error'
...
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string
  ReportingCenter:
    type: object
    properties:
      id:
        type: string
        description: Id of reporting center
      context:
        type: string
        description: Context of reporting center
      active:
        type: boolean
        description: Active center
      dateCreated:
        type: string
        format: date-time
        description: Creation date
  Sample:
    type: object
    properties:
      id:
        type: string
        description: Id of sample
      centerCode:
        type: string
        description: Center code
      active:
        type: boolean
        description: Active record
      dateCreated:
        type: string
        format: date-time
        description: Date record was created
      collectionMethods:
        type: array
        description: array of center codes
        items:
          $ref: '#/definitions/CollectionMethod'
...
  CollectionMethod:
    type: object
    properties:
      id:
        type: string
        description: Id of collection method
      name:
        type: string
        description: Name of method
      description:
        type: string
        description: Description of method
      active:
        type: boolean
        description: Active item
      dateCreated:
        type: string
        format: date-time
        description: Date entry was created
  TypingTestName:
    type: object
    properties:
      id:
        type: string
        description: Id of testing type name
      name:
        type: string
        description: Display / Common name of typing test
      description:
        type: string
        description: Description of typing test name
      active:
        type: boolean
        description: Active record
      dateCreated:
        type: string
        format: date-time
        description: Creation date
  Property:
    type: object
    properties:
      id:
        type: string
        description: Id of property
      name:
        type: string
        description: Display/ Common name of property
      value:
        type: string
        description: Value of property name
      description:
        type: string
        description: Description of property
      active:
        type: boolean
        description: Active record
      dateCreated:
        type: string
        format: date-time
        description: Creation date
      extendedItems:
        type: array
        description: Array of generic items
        items:
          $ref: '#/definitions/ExtendedItem'
  ExtendedItem:
    type: object
    properties:
      id:
        type: string
        description: Id of the extended item
      active:
        type: boolean
        description: Active record
      dateCreated:
        type: string
        format: date-time
        description: Creation date
      item:
        type: object
        description: Generic object
  Project:
    type: object
    properties:
      id:
        type: string
        description: Id of the project item
      name:
        type: string
        description: Name of project
      active:
        type: boolean
        description: Active record
      dateCreated:
        type: string
        format: date-time
        description: Creation date
  HmlId:
    type: object
    properties:
      id:
        type: string
        description: Id of the hmlid item
      rootName:
        type: string
        description: Root name of hml
      extension:
        type: string
        description: Extendsion of hml id
      active:
        type: boolean
        description: Active record
      dateCreated:
        type: string
        format: date-time
        description: Creation date
  Hml:
    type: object
    properties:
      id:
        type: string
        description: Id of the hml item
      version:
        type:
          $ref: '#/definitions/Version'
        description: Version of Hml
      project:
        type:
          $ref: '#/definitions/Project'
        description: Project object
      reportingCenters:
        type: array
        description: Repoting centers for this item
        items:
          $ref: '#/definitions/ReportingCenter'
      typingTestNames:
        type: array
        description: Typing test names used for this item
        items:
          $ref: '#/definitions/TypingTestName'
      samples:
        type: array
        description: Samples used in this item
        items:
          $ref: '#/definitions/Sample'
      properties:
        type: array
        description: Properteis belonging to this item
        items:
          $ref: '#/definitions/Property'
      active:
        type: boolean
        description: Active record
      dateCreated:
        type: string
        format: date-time
        description: Creation date
  Version:
    type: object
    properties:
      id:
        type: string
        description: Id of the version item
      name:
        type: string
        description: Name of the item
      description:
        type: string
        description: Short description of version
      active:
        type: boolean
        description: Active record
      dateCreated:
        type: string
        format: date-time
        description: Creation date